migrate/requirements.txt
beautifulsoup4
html2text
python-dateutil
click
aiohttp
